hack to allow loading school data
the school register contains two fields with cardinality 'n' --
`denominations` and `dioceses`.  This means that when loading from a
.tsv file, we need to know which columns are array columns and which are
not, so we can't just rely on the header line to tell us what is going
on.

this hack just hardcodes the columns in place.  in future, this could be
derived from register register / field register metadata.